requirements.txt
python-redmine~=2.4.0
python-dotenv~=0.21.1
